[fix] 피드백 받은 설정파일 위치 변경 및 예외 내부로 변경
