+++ src/main/java/com/prgrms/ohouse/web/commerce/requests/CartCreateRequest.java
@@ -1,7 +1,10 @@
 package com.prgrms.ohouse.web.commerce.requests;
 
+import lombok.AccessLevel;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
 public class CartCreateRequest {
 	private Long productId;
